+++ src/app/app-routing.module.ts
@@ -12,7 +12,8 @@ const routes: Routes = [
   },
   {
     path: 'login',
-    loadChildren: () => import('./auth/login.module').then(m => m.LoginModule),
+    loadChildren: () =>
+      import('./auth/login/login.module').then(m => m.LoginModule),
   },
   {
     path: 'main/:id',
