+++ src/app/app.module.ts
@@ -2,10 +2,9 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CoreModule } from './core/core.module';
 import { AppRoutingModule } from './app-routing.module';
-import { MainModule } from './youtube/main/main.module';
 
 @NgModule({
   declarations: [AppComponent],
@@ -14,6 +13,7 @@ import { MainModule } from './youtube/main/main.module';
     BrowserAnimationsModule,
     FormsModule,
     CoreModule,
+    ReactiveFormsModule,
     AppRoutingModule,
   ],
   providers: [],
