--- src/app/auth/login.component.html
@@ -1,3 +1,0 @@
-<p>login works!</p>
-<button class="btn btn-primary" (click)="logIn()">Login</button>
-<button class="btn btn-danger" (click)="logOut()">Login</button>
