--- src/app/auth/login.component.ts
@@ -1,19 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { AuthService } from './auth.service';
-
-@Component({
-  selector: 'app-login',
-  templateUrl: './login.component.html',
-  styleUrls: ['./login.component.scss'],
-})
-export class LoginComponent {
-  constructor(private authService: AuthService) {}
-
-  logIn() {
-    this.authService.login().subscribe();
-  }
-
-  logOut() {
-    this.authService.logout();
-  }
-}
