+++ src/app/auth/login/login.component.html
@@ -1,0 +1,29 @@
+<div class="d-flex align-items-center justify-content-center">
+  <form [formGroup]="login" (ngSubmit)="onSubmit()">
+    <h2>Login</h2>
+    <div class="form-group">
+      <label for="inputName" class="text-muted">Name</label>
+      <input
+        type="text"
+        class="form-control"
+        id="inputName"
+        formControlName="userName"
+        placeholder="Enter user name" />
+    </div>
+    <div class="form-group">
+      <label for="inputPassword" class="text-muted">Password</label>
+      <input
+        formControlName="password"
+        type="password"
+        class="form-control"
+        id="inputPassword"
+        placeholder="Password" />
+    </div>
+    <button class="btn btn-primary" type="submit" [disabled]="!login.valid">
+      Submit
+    </button>
+    <button class="btn btn-danger" type="button" (click)="logOut()">
+      Logout
+    </button>
+  </form>
+</div>
